Extract LookupJoiner class
